backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.3.0,<3.0.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.9
aiofiles==23.2.1
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.1
pypdf==4.2.0
python-pptx==0.6.23
Pillow==10.3.0
pytesseract==0.3.10
# LangChain ecosystem - use 0.3.x which supports google-generativeai 0.8+
langchain>=0.3.7,<0.4.0
langchain-google-genai>=2.0.5,<3.0.0
langchain-openai>=0.2.0,<1.0.0
openai>=1.40.0,<2.0.0
langchain-community>=0.3.7,<0.4.0
langgraph>=0.2.28,<0.3.0
langchain-core>=0.3.15,<0.4.0
google-generativeai>=0.8.0
chromadb>=1.0.0
sentence-transformers==2.7.0
python-dotenv==1.0.1
structlog==24.1.0
reportlab==4.2.0
openpyxl==3.1.2
pandas==2.2.2
chardet==5.2.0
zipfile36==0.1.3
